pyproject.toml
[project]
name = "imagetopdf"
version = "0.1.0"
description = ""
authors = [
    {name = "SMH642800",email = "SMH642800@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "pyqt6 (>=6.9.0,<7.0.0)",
    "img2pdf (>=0.6.1,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
